added render_math.js, show parse error from katex
